# shared/api/openapi.yaml
openapi: 3.0.3
info:
  title: Kubernetes Backup & Restore API
  description: |
    Comprehensive REST API for Kubernetes backup, restore, and disaster recovery operations.
    
    This API provides endpoints for:
    - Backup and restore operations
    - Disaster recovery scenario management
    - Cluster and resource validation
    - Integration bridge monitoring
    - Webhook management for GitOps workflows
    
    ## Authentication
    All API endpoints require appropriate authentication headers and role-based access control.
    
    ## Rate Limiting
    API endpoints are rate-limited to prevent abuse. Check response headers for current limits.
    
    ## Circuit Breaker
    External service calls are protected by circuit breakers for enhanced resilience.
  version: "1.0.0"
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.backup-system.example.com
    description: Production server
paths:
...
  /api/v1/restore:
    post:
      summary: Start restore operation
      description: Initiates a new restore operation from a backup
      operationId: startRestore
      tags:
        - Restore
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestoreRequest'
      responses:
        '202':
          description: Restore operation started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestoreOperationResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    get:
      summary: List active restores
      description: Returns a list of all currently active restore operations
      operationId: listActiveRestores
      tags:
        - Restore
      parameters:
        - name: limit
          in: query
          description: Maximum number of results to return
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Number of results to skip
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Active restores retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestoreListResponse'
  /api/v1/restore/{restoreId}:
    get:
      summary: Get restore status
      description: Returns the current status and progress of a specific restore operation
      operationId: getRestoreStatus
      tags:
        - Restore
      parameters:
        - name: restoreId
          in: path
          required: true
          description: Unique identifier of the restore operation
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Restore status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestoreOperationResponse'
        '404':
          description: Restore operation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    delete:
      summary: Cancel restore operation
      description: Cancels an active restore operation
      operationId: cancelRestore
      tags:
        - Restore
      parameters:
        - name: restoreId
          in: path
          required: true
          description: Unique identifier of the restore operation
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Restore operation cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Restore operation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to cancel restore operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/restore/history:
    get:
      summary: Get restore history
      description: Returns historical information about completed restore operations
      operationId: getRestoreHistory
      tags:
        - Restore
      parameters:
        - name: limit
          in: query
          description: Maximum number of results to return
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Number of results to skip
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: status
          in: query
          description: Filter by restore status
          schema:
            $ref: '#/components/schemas/RestoreStatus'
        - name: cluster_name
          in: query
          description: Filter by cluster name
          schema:
            type: string
      responses:
        '200':
          description: Restore history retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestoreHistoryResponse'
  /api/v1/restore/validate:
    post:
      summary: Validate restore request
      description: Validates a restore request without executing it
      operationId: validateRestore
      tags:
        - Restore
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestoreRequest'
      responses:
        '200':
          description: Validation completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResponse'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResponse'
  /api/v1/restore/plan:
    post:
      summary: Create restore plan
      description: Creates a detailed execution plan for a restore operation
      operationId: createRestorePlan
      tags:
        - Restore
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestoreRequest'
      responses:
        '200':
          description: Restore plan created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestorePlanResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Disaster Recovery Endpoints
...
  /api/v1/dr/execute:
    post:
      summary: Execute DR scenario
      description: Executes a disaster recovery scenario
      operationId: executeDRScenario
      tags:
        - Disaster Recovery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DisasterRecoveryRequest'
      responses:
        '202':
          description: DR scenario execution started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DRExecutionResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/dr/scenarios/{scenarioId}:
    get:
      summary: Get DR scenario status
      description: Returns the status of a disaster recovery scenario execution
      operationId: getDRScenarioStatus
      tags:
        - Disaster Recovery
      parameters:
        - name: scenarioId
          in: path
          required: true
          description: Unique identifier of the DR scenario
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: DR scenario status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DRExecutionResponse'
        '404':
          description: DR scenario not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Backup Management Endpoints
  /api/v1/backups:
    get:
      summary: List available backups
      description: Returns a list of available backups for restore operations
      operationId: listAvailableBackups
      tags:
        - Backups
      parameters:
        - name: cluster_name
          in: query
          description: Filter by cluster name
          schema:
            type: string
        - name: namespace
          in: query
          description: Filter by namespace
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of results to return
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Available backups retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackupListResponse'
  /api/v1/backups/{backupId}:
    get:
      summary: Get backup details
      description: Returns detailed information about a specific backup
      operationId: getBackupDetails
      tags:
        - Backups
      parameters:
        - name: backupId
          in: path
          required: true
          description: Unique identifier of the backup
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Backup details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackupDetailsResponse'
        '404':
          description: Backup not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/backups/{backupId}/validate:
    post:
      summary: Validate backup
      description: Validates the integrity and completeness of a backup
      operationId: validateBackup
      tags:
        - Backups
      parameters:
        - name: backupId
          in: path
          required: true
          description: Unique identifier of the backup
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Backup validation completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResponse'
        '404':
          description: Backup not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Cluster Management Endpoints
...
  /api/v1/clusters/{clusterName}/validate:
    post:
      summary: Validate cluster
      description: Validates cluster configuration and connectivity
      operationId: validateCluster
      tags:
        - Clusters
      parameters:
        - name: clusterName
          in: path
          required: true
          description: Name of the cluster to validate
          schema:
            type: string
      responses:
        '200':
          description: Cluster validation completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResponse'
        '404':
          description: Cluster not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    RestoreStatus:
      type: string
      enum: [pending, validating, restoring, completed, failed, cancelled]
      description: Current status of the restore operation
      example: "restoring"
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
    
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
    
    BasicAuth:
      type: http
      scheme: basic
      description: Basic HTTP authentication
security:
  - ApiKeyAuth: []
  - BearerAuth: []
  - BasicAuth: []
...